Buff up reminders

+++ app/src/main/java/com/example/vetapp/viewmodels/RemindersViewModel.kt
@@ -1,17 +1,22 @@
 package com.example.vetapp.viewmodels
 
+import android.Manifest
 import android.app.AlarmManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
+import android.content.pm.PackageManager
 import android.os.Build
 import android.util.Log
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.vetapp.Database.Entities.Reminder
+import com.example.vetapp.MainActivity
 import com.example.vetapp.VetApplication
-import com.example.vetapp.broadcastreceivers.FillOutReportsNotificationReceiver
-import com.example.vetapp.broadcastreceivers.SendReportsNotificationReceiver
+import com.example.vetapp.broadcastreceivers.NotificationReceiver
+import com.example.vetapp.reminders.RequestCodeFactory
 import com.example.vetapp.repositories.IRemindersRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
@@ -40,25 +45,21 @@ class RemindersViewModel @Inject constructor(
     var isError : StateFlow<Boolean> = _isError
     var errorMsg : StateFlow<String> = _errorMsg
 
-    public fun scheduleReminder(date : String, time : String, recurrence : String, type : String){
-        val context = VetApplication.applicationContext()
-        val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-        var intent : Intent
-        when(type){
-            "Send" -> intent = Intent(context, SendReportsNotificationReceiver::class.java)
-            "Fill" -> intent = Intent(context, FillOutReportsNotificationReceiver::class.java)
-            else -> return
-        }
-        val pendingIntent = PendingIntent.getBroadcast(context, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE)
+    public fun scheduleReminder(date : String, time : String, recurrence : String, title: String, message : String){
+        Log.d(TAG, "Scheduling reminder")
         val dateTimeFormat = SimpleDateFormat("dd/MM/yyyy HH:mm", Locale.getDefault())
         val calendar = Calendar.getInstance()
+        val context = VetApplication.applicationContext()
+        val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+        val requestCode = RequestCodeFactory.GetRequestCode()
+        val pendingIntent = BuildReminderPendingIntent(context, title, message, requestCode)
         try {
             // Combine the date and time into a single string for parsing
             val dateTimeString = "$date $time"
             val dateTime = dateTimeFormat.parse(dateTimeString)
             calendar.time = dateTime
         } catch (e: Exception) {
-            e.printStackTrace()
+            Log.e(TAG, "Exception scheduling reminder: ${e.printStackTrace()}")
             return
         }
         when (recurrence) {
@@ -67,9 +68,13 @@ class RemindersViewModel @Inject constructor(
                     if(alarmManager.canScheduleExactAlarms()){
                         alarmManager.setExact(AlarmManager.RTC_WAKEUP, calendar.timeInMillis, pendingIntent)
                     }
+                    else{
+                        Log.e(TAG, "Dont have permissions to schedule exact alerms")
+                        return
+                    }
                 }
                 else{
-                    //TODO how to do it for old api versions
+                    Log.e(TAG, "Exception scheduling reminder:")
                     return
                 }
 
@@ -93,10 +98,18 @@ class RemindersViewModel @Inject constructor(
             // Add more recurrence types as needed
             else -> return
         }
-        Log.d(TAG, "Set a reminder for ${type}")
-        val reminder = Reminder(type = type, frequency = recurrence, startTime = time)
+        Log.d(TAG, "Set a reminder for ${title}")
+        val reminder = Reminder(type = title, frequency = recurrence, startTime = time, requestCode = requestCode, title = title, message = message)
         insertReminder(reminder)
         Log.d(TAG, "Saved reminder to database")
+    }
+
+    fun BuildReminderPendingIntent(context: Context, title : String, message : String, requestCode : Int) : PendingIntent{
+        var intent = Intent(context, NotificationReceiver::class.java)
+        intent.putExtra("notification_title", title)
+        intent.putExtra("notification_message", message)
+        val pendingIntent = PendingIntent.getBroadcast(context, requestCode, intent, PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE)
+        return pendingIntent
     }
 
     fun checkAndRequestExactAlarmPermission() : Boolean{
@@ -122,9 +135,35 @@ class RemindersViewModel @Inject constructor(
     fun deleteReminder(reminder: Reminder){
         viewModelScope.launch {
             withContext(Dispatchers.IO){
-                val result = remindersRepository.deleteReminder(reminder)
+                var result = cancelAlarmForReminder(reminder)
+                if(!result){
+                    return@withContext
+                }
+                try{
+                    remindersRepository.deleteReminder(reminder)
+                }
+                catch(e : Exception){
+                    Log.e(TAG, "Failed to delete notification from database ${e.message}")
+                }
             }
         }
+    }
+
+    fun cancelAlarmForReminder(reminder: Reminder) : Boolean{
+        try{
+            val context = VetApplication.applicationContext()
+            val intent = Intent(context, NotificationReceiver::class.java)
+            intent.putExtra("notification_title", reminder.title)
+            intent.putExtra("notification_message", reminder.message)
+            val pendingIntent = PendingIntent.getBroadcast(context, reminder.requestCode, intent, PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE)
+            val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+            alarmManager.cancel(pendingIntent)
+        }
+        catch(e : Exception){
+            Log.e(TAG, "Failed to delete alarm for notification ${e.message}")
+            return false
+        }
+        return true
     }
 
     fun UpdateErrorState(isError : Boolean, errorMsg : String){
